+++ docs/CHANGELOG.md
@@ -10,6 +10,9 @@ Jupytext ChangeLog
 **Added**
 - Added Haskell as supported language ([#909](https://github.com/mwouts/jupytext/issues/909)) - thanks to [codeweber](https://github.com/codeweber) for this contribution
 
+**Changed**
+- We have updated the pre-commit hooks and in particular we switched to the first stable version of `black==22.1.0`.
+
 
 1.13.6 (2022-01-11)
 -------------------